+++ swf/models/workflow.py
@@ -173,10 +173,10 @@ class WorkflowType(BaseModel):
     @property
     @exceptions.translate(SWFResponseError, to=ResponseError)
     @exceptions.is_not(WorkflowTypeDoesNotExist)
-    @exceptions.when(SWFResponseError,
-                     raises(WorkflowTypeDoesNotExist,
-                            when=exceptions.is_unknown('WorkflowType'),
-                            extract=exceptions.extract_resource))
+    @exceptions.catch(SWFResponseError,
+                      raises(WorkflowTypeDoesNotExist,
+                             when=exceptions.is_unknown('WorkflowType'),
+                             extract=exceptions.extract_resource))
     def exists(self):
         """Checks if the WorkflowType exists amazon-side
 
@@ -246,7 +246,7 @@ class WorkflowType(BaseModel):
         :type   input: dict
 
         :param  tag_list: Tags associated with the workflow execution
-        :type   tag_list: String
+        :type   tag_list: String or list of strings
 
         :param  decision_tasks_timeout: maximum duration of decision tasks
                                         for this workflow execution
@@ -256,6 +256,11 @@ class WorkflowType(BaseModel):
         task_list = task_list or self.task_list
         child_policy = child_policy or self.child_policy
         input = json.dumps(input) or None
+        tag_list = tag_list if isinstance(tag_list, list) else [tag_list]
+
+        # checks
+        if len(tag_list) > 5:
+            raise ValueError("You cannot have more than 5 tags in StartWorkflowExecution.")
 
         run_id = self.connection.start_workflow_execution(
             self.domain.name,
@@ -420,10 +425,10 @@ class WorkflowExecution(BaseModel):
     @property
     @exceptions.translate(SWFResponseError, to=ResponseError)
     @exceptions.is_not(WorkflowExecutionDoesNotExist)
-    @exceptions.when(SWFResponseError,
-                     raises(WorkflowExecutionDoesNotExist,
-                            when=exceptions.is_unknown('WorkflowExecution'),
-                            extract=exceptions.extract_resource))
+    @exceptions.catch(SWFResponseError,
+                      raises(WorkflowExecutionDoesNotExist,
+                             when=exceptions.is_unknown('WorkflowExecution'),
+                             extract=exceptions.extract_resource))
     def exists(self):
         """Checks if the WorkflowExecution exists amazon-side
 
@@ -476,10 +481,10 @@ class WorkflowExecution(BaseModel):
 
     @exceptions.translate(SWFResponseError,
                           to=ResponseError)
-    @exceptions.when(SWFResponseError,
-                     raises(WorkflowExecutionDoesNotExist,
-                            when=exceptions.is_unknown('WorkflowExecution'),
-                            extract=exceptions.extract_resource))
+    @exceptions.catch(SWFResponseError,
+                      raises(WorkflowExecutionDoesNotExist,
+                             when=exceptions.is_unknown('WorkflowExecution'),
+                             extract=exceptions.extract_resource))
     def signal(self, signal_name, input=None, *args, **kwargs):
         """Records a signal event in the workflow execution history and
         creates a decision task.
@@ -506,10 +511,10 @@ class WorkflowExecution(BaseModel):
 
     @exceptions.translate(SWFResponseError,
                           to=ResponseError)
-    @exceptions.when(SWFResponseError,
-                     raises(WorkflowExecutionDoesNotExist,
-                            when=exceptions.is_unknown('domain'),
-                            extract=exceptions.extract_resource))
+    @exceptions.catch(SWFResponseError,
+                      raises(WorkflowExecutionDoesNotExist,
+                             when=exceptions.is_unknown('domain'),
+                             extract=exceptions.extract_resource))
     def request_cancel(self, *args, **kwargs):
         """Requests the workflow execution cancel"""
         self.connection.request_cancel_workflow_execution(
@@ -519,10 +524,10 @@ class WorkflowExecution(BaseModel):
 
     @exceptions.translate(SWFResponseError,
                           to=ResponseError)
-    @exceptions.when(SWFResponseError,
-                     raises(WorkflowExecutionDoesNotExist,
-                            when=exceptions.is_unknown('domain'),
-                            extract=exceptions.extract_resource))
+    @exceptions.catch(SWFResponseError,
+                      raises(WorkflowExecutionDoesNotExist,
+                             when=exceptions.is_unknown('domain'),
+                             extract=exceptions.extract_resource))
     def terminate(self, *args, **kwargs):
         """Terminates the workflow execution"""
         self.connection.terminate_workflow_execution(