+++ HISTORY.rst
@@ -1,6 +1,24 @@
 Changelog
 ---------
 
+0.11.12
+~~~~~~~
+
+- Fix --tags argument for simpleflow standalone (#114)
+- Improve tests and add integration tests (#116)
+- Add 'simpleflow activity.rerun' command (#117)
+
+0.11.11
+~~~~~~~
+
+- Fix a circular import on simpleflow.swf.executor
+
+0.11.10
+~~~~~~~
+
+- Fix previous_history initialization (#106)
+- Improve WorkflowExecutionQueryset default date values (#111)
+
 0.11.9
 ~~~~~~